Add LiquidHaskell plugin pragma to every source file
This is needed because linking seems to fail whenever the plugin is configured
in the `package.yaml` file, so it needs to be configured in the source files.

Still, some verifications fail when the plugin is not configured in
`package.yaml`, so the workflow for now is to compile WITH it there (but linking
fails) and after that to link WITHOUT it.