Try to improve fix for CSE bug in equality assertion
The solution from the previous commit `1890897` still creates more wires than
necessary and leaves one of them unused. This commit attempts to fix that by
adding one more argument to `label'` that requests to use one particular wire
for labeling the output of a program, but a proof is still missing.